reorganization of resolver tests
